requirements.txt
python-vlc>=3.0.0
# Note: tkinter is included with the Python standard library on most platforms
# The VLC native library (libVLC) must be installed separately on your system.
